feat(MetricOptions): reusable component for Metric options such as range
